# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/system/miniapp/employee/overview": {
      "get": {
        "operationId": "SystemMiniappEmployeeController_overview",
        "parameters": [
          {
            "name": "range",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "员工-基础运营概览（今日/近7/近30/本月/上月）",
        "tags": [
          "MiniappEmployee"
        ]
      }
    },
    "/system/miniapp/employee/daily": {
      "get": {
        "operationId": "SystemMiniappEmployeeController_daily",
        "parameters": [
          {
            "name": "range",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "员工-按日指标列表（净支付/洗车卡划扣/洗车数量合计）",
        "tags": [
          "MiniappEmployee"
        ]
      }
    }
  },
  "info": {
    "title": "Wash Club API",
    "description": "API for 一体化洗车门店管理平台",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}